Move EntityType and related types to postgres-core-model
