+++ crates/postgres-subsystem/postgres-core-model/src/lib.rs
@@ -1,3 +1,7 @@
 pub mod access;
+pub mod aggregate;
 pub mod migration;
+pub mod relation;
 pub mod subsystem;
+pub mod types;
+pub mod vector_distance;
