+++ crates/postgres-subsystem/postgres-graphql-builder/src/access_utils.rs
@@ -33,11 +33,12 @@ use core_plugin_interface::{
 };
 
 use exo_sql::{ColumnPathLink, Database, PhysicalColumnPath};
-use postgres_core_model::access::{
-    DatabaseAccessPrimitiveExpression, InputAccessPrimitiveExpression,
+use postgres_core_model::types::{
+    base_type, EntityType, PostgresFieldType, PostgresPrimitiveType, PostgresType,
 };
-use postgres_graphql_model::types::{
-    base_type, EntityType, PostgresField, PostgresFieldType, PostgresPrimitiveType, PostgresType,
+use postgres_core_model::{
+    access::{DatabaseAccessPrimitiveExpression, InputAccessPrimitiveExpression},
+    types::PostgresField,
 };
 
 use postgres_core_builder::resolved_type::ResolvedTypeEnv;
