+++ crates/postgres-subsystem/postgres-graphql-builder/src/delete_mutation_builder.rs
@@ -18,8 +18,8 @@ use core_plugin_interface::{
     },
     core_model_builder::error::ModelBuildingError,
 };
+use postgres_core_model::types::EntityType;
 use postgres_graphql_model::mutation::PostgresMutationParameters;
-use postgres_graphql_model::types::EntityType;
 
 use super::{
     builder::Builder, mutation_builder::MutationBuilder, naming::ToPostgresMutationNames,
