+++ crates/postgres-subsystem/postgres-graphql-builder/src/mutation_builder.rs
@@ -21,14 +21,14 @@ use core_plugin_interface::{
 
 use postgres_graphql_model::{
     mutation::{PostgresMutation, PostgresMutationParameters},
-    relation::PostgresRelation,
-    types::{
-        base_type, EntityType, MutationType, PostgresField, PostgresFieldType, PostgresType,
-        TypeIndex,
-    },
+    types::MutationType,
 };
 
-use postgres_core_model::access::DatabaseAccessPrimitiveExpression;
+use postgres_core_model::{
+    access::DatabaseAccessPrimitiveExpression,
+    relation::PostgresRelation,
+    types::{base_type, EntityType, PostgresField, PostgresFieldType, PostgresType, TypeIndex},
+};
 
 use crate::{access_utils::parent_predicate, shallow::Shallow, utils::to_mutation_type};
