+++ crates/postgres-subsystem/postgres-graphql-builder/src/order_by_type_builder.rs
@@ -16,8 +16,9 @@ use exo_sql::{ColumnPathLink, Database};
 use postgres_graphql_model::{
     order::OrderByParameter,
     order::{OrderByParameterType, OrderByParameterTypeKind, OrderByParameterTypeWrapper},
-    types::{EntityType, PostgresField, PostgresPrimitiveType, PostgresType},
 };
+
+use postgres_core_model::types::{EntityType, PostgresField, PostgresPrimitiveType, PostgresType};
 
 use postgres_core_model::access::Access;
