+++ crates/postgres-subsystem/postgres-graphql-builder/src/type_builder.rs
@@ -31,7 +31,11 @@ use core_plugin_interface::{
 
 use exo_sql::{ColumnId, VectorDistanceFunction, DEFAULT_VECTOR_SIZE};
 
-use postgres_graphql_model::{
+use postgres_core_model::{
+    access::{
+        Access, DatabaseAccessPrimitiveExpression, InputAccessPrimitiveExpression,
+        UpdateAccessExpression,
+    },
     aggregate::{AggregateField, AggregateFieldType},
     relation::{ManyToOneRelation, OneToManyRelation, PostgresRelation, RelationCardinality},
     types::{
@@ -39,11 +43,6 @@ use postgres_graphql_model::{
         TypeIndex,
     },
     vector_distance::{VectorDistanceField, VectorDistanceType},
-};
-
-use postgres_core_model::access::{
-    Access, DatabaseAccessPrimitiveExpression, InputAccessPrimitiveExpression,
-    UpdateAccessExpression,
 };
 
 use super::naming::ToPostgresQueryName;
