+++ crates/postgres-subsystem/postgres-graphql-builder/src/utils.rs
@@ -8,7 +8,8 @@
 // by the Apache License, Version 2.0.
 
 use core_plugin_interface::core_model::types::FieldType;
-use postgres_graphql_model::types::{EntityType, MutationType, PostgresFieldType, TypeIndex};
+use postgres_core_model::types::{EntityType, PostgresFieldType, TypeIndex};
+use postgres_graphql_model::types::MutationType;
 
 pub(super) fn to_mutation_type(
     field_type: &FieldType<PostgresFieldType<EntityType>>,
