+++ crates/postgres-subsystem/postgres-graphql-model/src/types.rs
@@ -7,123 +7,32 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use super::relation::PostgresRelation;
-use crate::aggregate::AggregateField;
+use crate::operation::{OperationParameters, PostgresOperation};
 use crate::query::CollectionQueryParameters;
-use crate::relation::OneToManyRelation;
 use crate::subsystem::PostgresGraphQLSubsystem;
-use crate::vector_distance::VectorDistanceField;
 use async_graphql_parser::types::{
     FieldDefinition, InputObjectType, ObjectType, Type, TypeDefinition, TypeKind,
 };
 use core_plugin_interface::core_model::access::AccessPredicateExpression;
-use core_plugin_interface::core_model::context_type::ContextSelection;
 use core_plugin_interface::core_model::primitive_type::vector_introspection_base_type;
 use core_plugin_interface::core_model::types::{DirectivesProvider, TypeValidation};
 use core_plugin_interface::core_model::{
-    mapped_arena::{SerializableSlab, SerializableSlabIndex},
+    mapped_arena::SerializableSlabIndex,
     type_normalization::{
         default_positioned, default_positioned_name, FieldDefinitionProvider, InputValueProvider,
         Parameter, TypeDefinitionProvider,
     },
     types::{FieldType, Named},
 };
+use postgres_core_model::relation::OneToManyRelation;
+use postgres_core_model::relation::PostgresRelation;
 
 use postgres_core_model::access::{
-    Access, DatabaseAccessPrimitiveExpression, InputAccessPrimitiveExpression,
+    DatabaseAccessPrimitiveExpression, InputAccessPrimitiveExpression,
 };
 
-use exo_sql::PhysicalTable;
+use postgres_core_model::types::{EntityType, PostgresField, PostgresPrimitiveType};
 use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug, Clone, Copy)]
-pub enum TypeIndex<CT> {
-    Primitive(SerializableSlabIndex<PostgresPrimitiveType>),
-    Composite(SerializableSlabIndex<CT>),
-}
-
-impl<CT> TypeIndex<CT> {
-    pub fn to_type<'a>(
-        &self,
-        primitive_types: &'a SerializableSlab<PostgresPrimitiveType>,
-        entity_types: &'a SerializableSlab<CT>,
-    ) -> PostgresType<'a, CT> {
-        match self {
-            TypeIndex::Primitive(index) => PostgresType::Primitive(&primitive_types[*index]),
-            TypeIndex::Composite(index) => PostgresType::Composite(&entity_types[*index]),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub enum PostgresType<'a, CT> {
-    Primitive(&'a PostgresPrimitiveType),
-    Composite(&'a CT),
-}
-
-impl<'a, CT: Named> PostgresType<'a, CT> {
-    pub fn name(&self) -> &str {
-        match self {
-            PostgresType::Primitive(primitive_type) => &primitive_type.name,
-            PostgresType::Composite(composite_type) => composite_type.name(),
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct PostgresPrimitiveType {
-    pub name: String,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct EntityType {
-    pub name: String,
-    pub plural_name: String,
-
-    pub fields: Vec<PostgresField<EntityType>>,
-    pub agg_fields: Vec<AggregateField>,
-    pub vector_distance_fields: Vec<VectorDistanceField>,
-
-    pub table_id: SerializableSlabIndex<PhysicalTable>,
-    pub access: Access,
-}
-
-pub fn get_field_id(
-    types: &SerializableSlab<EntityType>,
-    entity_id: SerializableSlabIndex<EntityType>,
-    field_name: &str,
-) -> Option<EntityFieldId> {
-    let entity = &types[entity_id];
-    entity
-        .fields
-        .iter()
-        .position(|field| field.name == field_name)
-        .map(|field_index| EntityFieldId(field_index, entity_id))
-}
-
-/// Encapsulates a field on an entity type (mirros how `ColumnId` is structured)
-#[derive(Serialize, Deserialize, Clone, Copy, Debug)]
-pub struct EntityFieldId(usize, SerializableSlabIndex<EntityType>);
-
-impl EntityFieldId {
-    pub fn entity_type_id(&self) -> SerializableSlabIndex<EntityType> {
-        self.1
-    }
-
-    pub fn resolve<'a>(
-        &self,
-        types: &'a SerializableSlab<EntityType>,
-    ) -> &'a PostgresField<EntityType> {
-        let entity = &types[self.1];
-        &entity.fields[self.0]
-    }
-}
-
-impl Named for EntityType {
-    fn name(&self) -> &str {
-        &self.name
-    }
-}
 
 /// Mutation input types such as `CreatePostInput` and `UpdatePostInput`
 #[derive(Serialize, Deserialize, Debug, Clone)]
@@ -136,72 +45,6 @@ pub struct MutationType {
         Option<SerializableSlabIndex<AccessPredicateExpression<InputAccessPrimitiveExpression>>>,
     pub database_access:
         Option<SerializableSlabIndex<AccessPredicateExpression<DatabaseAccessPrimitiveExpression>>>,
-}
-
-impl EntityType {
-    pub fn field_by_name(&self, name: &str) -> Option<&PostgresField<EntityType>> {
-        self.fields.iter().find(|field| field.name == name)
-    }
-
-    pub fn pk_field(&self) -> Option<&PostgresField<EntityType>> {
-        self.fields
-            .iter()
-            .find(|field| matches!(&field.relation, PostgresRelation::Pk { .. }))
-    }
-
-    pub fn pk_field_id(
-        &self,
-        entity_id: SerializableSlabIndex<EntityType>,
-    ) -> Option<EntityFieldId> {
-        self.fields
-            .iter()
-            .position(|field| matches!(&field.relation, PostgresRelation::Pk { .. }))
-            .map(|field_index| EntityFieldId(field_index, entity_id))
-    }
-
-    pub fn aggregate_field_by_name(&self, name: &str) -> Option<&AggregateField> {
-        self.agg_fields.iter().find(|field| field.name == name)
-    }
-
-    pub fn vector_distance_field_by_name(&self, name: &str) -> Option<&VectorDistanceField> {
-        self.vector_distance_fields
-            .iter()
-            .find(|field| field.name == name)
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct PostgresField<CT> {
-    pub name: String,
-    pub typ: FieldType<PostgresFieldType<CT>>,
-    pub relation: PostgresRelation,
-    pub has_default_value: bool, // does this field have a default value?
-    pub dynamic_default_value: Option<ContextSelection>,
-    pub readonly: bool,
-    pub access: Access,
-    pub type_validation: Option<TypeValidation>,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct PostgresFieldType<CT> {
-    pub type_id: TypeIndex<CT>,
-    pub type_name: String,
-}
-
-impl<CT> Named for PostgresFieldType<CT> {
-    fn name(&self) -> &str {
-        &self.type_name
-    }
-}
-
-pub fn base_type<'a, CT>(
-    typ: &FieldType<PostgresFieldType<CT>>,
-    primitive_types: &'a SerializableSlab<PostgresPrimitiveType>,
-    entity_types: &'a SerializableSlab<CT>,
-) -> PostgresType<'a, CT> {
-    typ.innermost()
-        .type_id
-        .to_type(primitive_types, entity_types)
 }
 
 impl TypeDefinitionProvider<PostgresGraphQLSubsystem> for PostgresPrimitiveType {
@@ -265,7 +108,9 @@ impl TypeDefinitionProvider<PostgresGraphQLSubsystem> for MutationType {
                 .fields
                 .iter()
                 .flat_map(|field| {
-                    (!field.readonly).then_some(default_positioned(field.input_value()))
+                    (!field.readonly).then_some(default_positioned(
+                        PostgresMutationField(field).input_value(),
+                    ))
                 })
                 .collect();
             TypeKind::InputObject(InputObjectType { fields })
@@ -352,16 +197,43 @@ impl<CT> FieldDefinitionProvider<PostgresGraphQLSubsystem> for PostgresField<CT>
     }
 }
 
-impl<CT> Parameter for PostgresField<CT> {
+// To get around the orphan rule, we wrap the field in a struct
+struct PostgresMutationField<'a>(&'a PostgresField<MutationType>);
+
+impl<'a> Parameter for PostgresMutationField<'a> {
     fn name(&self) -> &str {
-        &self.name
+        &self.0.name
     }
 
     fn typ(&self) -> Type {
-        (&self.typ).into()
+        (&self.0.typ).into()
     }
 
     fn type_validation(&self) -> Option<TypeValidation> {
-        self.type_validation.clone()
+        self.0.type_validation.clone()
+    }
+}
+
+pub trait Operation {
+    fn name(&self) -> &String;
+    fn parameters(&self) -> Vec<&dyn Parameter>;
+    fn return_type(&self) -> Type;
+}
+
+impl<S, P: OperationParameters> FieldDefinitionProvider<S> for PostgresOperation<P> {
+    fn field_definition(&self, _system: &S) -> FieldDefinition {
+        let fields = self
+            .parameters()
+            .iter()
+            .map(|parameter| default_positioned(parameter.input_value()))
+            .collect();
+
+        FieldDefinition {
+            description: None,
+            name: default_positioned_name(self.name()),
+            arguments: fields,
+            directives: vec![],
+            ty: default_positioned(self.return_type()),
+        }
     }
 }
