+++ crates/postgres-subsystem/postgres-graphql-model/src/vector_distance.rs
@@ -7,47 +7,12 @@ use async_graphql_value::Name;
 use core_plugin_interface::core_model::type_normalization::{
     default_positioned, default_positioned_name, FieldDefinitionProvider, TypeDefinitionProvider,
 };
-use exo_sql::{ColumnId, VectorDistanceFunction};
-use serde::{Deserialize, Serialize};
 
 use crate::subsystem::PostgresGraphQLSubsystem;
 
-use postgres_core_model::access::Access;
-
-/// Field for a vector distance function
-/// Represents:
-/// ```graphql
-/// document {
-///    contentVector: [Float!]!
-///    contentVectorDistance(to: [Float!]!): Float! <--- This is the field
-/// }
-/// ```
-#[derive(Serialize, Deserialize, Debug)]
-pub struct VectorDistanceField {
-    pub name: String,
-    pub column_id: ColumnId,
-    pub size: usize,
-    pub distance_function: VectorDistanceFunction,
-    pub access: Access,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct VectorDistanceType {
-    pub name: String, // name of the type, currently always "VectorDistance", but we could introduce `VectorDistance64`, etc. in the future
-    fields: Vec<VectorDistanceTypeField>,
-}
-
-impl VectorDistanceType {
-    pub fn new(name: String) -> Self {
-        Self {
-            name,
-            fields: vec![VectorDistanceTypeField {}],
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct VectorDistanceTypeField {}
+use postgres_core_model::vector_distance::{
+    VectorDistanceField, VectorDistanceType, VectorDistanceTypeField,
+};
 
 /// The definition of a vector distance field. Takes and argument
 impl FieldDefinitionProvider<PostgresGraphQLSubsystem> for VectorDistanceField {
